requirements.txt
eth_keys
eth-account>=0.4.0,<0.6.0
eth-abi
uniswap-universal-router-decoder
web3>=5.0.0,<6.0.0
simple-multicall
responses
python-dotenv>=1.0.0